% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TMBpred.R
\name{assignClass}
\alias{assignClass}
\title{AssignClass}
\usage{
assignClass(x, prior, type = "low_high")
}
\arguments{
\item{x}{predicted TMB}

\item{prior}{prior parameters.}

\item{type}{If 'low_high", only class 2 and 3 will be grouped to high.
If 'exact', exact class will be reported}
}
\value{
A vector of class
}
\description{
predict TMB stat function
}
\examples{
\dontrun{
assignClass(Data)
}

}
