% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecTMB.R
\name{calMut}
\alias{calMut}
\title{calMut
provide observed and predicted mutation rate}
\usage{
calMut(Data, params, sampleN = NULL, gids = NULL, bed = NULL)
}
\arguments{
\item{Data}{MutSet class. Detail see \code{MutSet}.}

\item{params}{output from fit_model. Detail see \code{fit_model}}

\item{sampleN}{Sample names}

\item{gids}{gene IDs}

\item{bed}{path to bed file.}
}
\value{
a summary of predicted and expected mutation rate.
}
\description{
calMut
provide observed and predicted mutation rate
}
