% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{CalTMB}
\alias{CalTMB}
\title{CalTMB}
\usage{
CalTMB(x, type = "all", sampleN = NULL)
}
\arguments{
\item{x}{mutSet object}

\item{type}{Type of mutation to be included. It can be 'all' 'nonsil' 'sil' 'indel' 'snp' 'frameshit'}

\item{sampleN}{Default is NULL. If it is specified, only subset of samples will be reported.}
}
\value{
A data.frame contain the mutation count
}
\description{
calcuate number of mutation per patient
}
\examples{
\dontrun{
CalTMB(Data)
}

}
