% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{checkMutC}
\alias{checkMutC}
\title{check each sample have at least one mutation.}
\usage{
checkMutC(sample, mut)
}
\arguments{
\item{sample}{data.frame of sample}

\item{mut}{Mutation table}
}
\value{
restrict sample to only the ones that contain at least one mutation
}
\description{
check each sample have at least one mutation.
}
