% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TMBpred.R
\name{CM}
\alias{CM}
\title{CM}
\usage{
CM(Data, mu, o_scale, MRtriProb, gids, zero_p = 0)
}
\arguments{
\item{Data}{mutSet}

\item{mu}{mu of the gene}

\item{o_scale}{o_scale}

\item{MRtriProb}{MRtriProb}

\item{gids}{gene IDs}

\item{zero_p}{zero_p}
}
\value{
A list for all the inputs for Instan
}
\description{
predict number of mutation
}
\examples{
\dontrun{
CM(Data)
}

}
