% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/procMaf.R
\name{ConvertExtraType}
\alias{ConvertExtraType}
\title{ConvertExtraType}
\usage{
ConvertExtraType(maf)
}
\arguments{
\item{maf}{maf matrix}
}
\value{
maf file with converted extra type
}
\description{
ConvertExtraType
}
\examples{
\dontrun{
ConvertExtraType(maf)
}

}
