% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{count_Mut}
\alias{count_Mut}
\title{Count mutatin per patient or per mutation tri-nuclear context.}
\usage{
count_Mut(mut, gid = NULL, sampleN = NULL, byContext = FALSE)
}
\arguments{
\item{mut}{data frame contain mutation info which must contain 'Start_Position' and 'Tumor_Sample_Barcode' or
exomeGene - data.frame.}

\item{gid}{default is NULL. provide gid list}

\item{sampleN}{default is NULL. Provide list of sample name. This parameter is only useful when byContext is FALSE}

\item{byContext}{A boolen. If TRUE, the number of mutation per tri-nucleotide context will be reported.
If FALSE, the number of mutation per patient will be reported. Default is FALSE.}
}
\value{
mutCount data.frame contain number of mutation for each patient or
a vector contains number of mutation per tri-nucleotide context
}
\description{
Count mutatin per patient or per mutation tri-nuclear context.
}
\examples{
\dontrun{
count_Mut(mut)
}
}
