% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TMBpred.R
\name{ExGMM}
\alias{ExGMM}
\title{ExGMM}
\usage{
ExGMM(x, msi = NULL, single = FALSE)
}
\arguments{
\item{x}{number of mutation}

\item{msi}{msi status}

\item{single}{A boolen. If True, only one class will be identified.}
}
\value{
A vector mutation count for 96 trinucleotide context
}
\description{
fit the data distribution to a mixture of Gaussian mixture
}
\examples{
\dontrun{
ExGMM(Data)
}

}
