% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{fit_model}
\alias{fit_model}
\title{fit_model}
\usage{
fit_model(Data, MRtriProb, method = "NB", cores = 1, bs.type = "all",
  mut.nonsil = FALSE, nonsil.fraction = 0.6)
}
\arguments{
\item{Data}{mutSet object}

\item{MRtriProb}{output from get_bg_MRtri}

\item{method}{Either NB or Poisson}

\item{cores}{Number of parallel core to be used}

\item{bs.type}{Either 'all' or nonsil. When all is specificed, all mutation per Mb
will be used as sample specific background mutation (TMB). When nonsil is specified,
non-synonymous mutation per Mb will be used as sample specific background mutation (TMB).}

\item{mut.nonsil}{A boolen. If TRUE, non-synonymous mutation will be used for background mutation
modeling during training.}

\item{nonsil.fraction}{Default is 0.6. The percentage of genes whose non-synonymous mutation
will be used.}
}
\value{
A vector mutation count for 96 trinucleotide context
}
\description{
fit background model with negative-binomial regression
}
\examples{
\dontrun{
fit_model2(Data)
}

}
