% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_exomeGene}
\alias{get_exomeGene}
\title{Get exomeGene}
\usage{
get_exomeGene(exome, Bed = NULL, mutContext)
}
\arguments{
\item{exome}{exome data}

\item{Bed}{path to bed file or data.frame}

\item{mutContext}{data.frame contain mutation context.}
}
\value{
exomeGene file
}
\description{
Get exomeGene
}
\examples{
\dontrun{
get_exomeGene(exome, Bed, mutContext)
}

}
