% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{get_genewise_dispersion_mle_mu}
\alias{get_genewise_dispersion_mle_mu}
\title{MLE to estimate genewise dispersion ###}
\usage{
get_genewise_dispersion_mle_mu(y, mu, offset, cores, ZI = FALSE)
}
\arguments{
\item{y}{input y}

\item{mu}{mu}

\item{offset}{offset}

\item{cores}{number of threads}

\item{ZI}{A boolen. Default is FALSE which Negative binomial model will be used.
If TRUE, zero-inflated model will be used.}
}
\value{
mu
}
\description{
MLE to estimate genewise dispersion ###
}
