% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_glen}
\alias{get_glen}
\title{get_glen}
\usage{
get_glen(exomeGene, selGid = NULL, byGene = FALSE)
}
\arguments{
\item{exomeGene}{data.frame}

\item{selGid}{a vector of selected genes.}

\item{byGene}{a boolen.If TRUE, a vector of length for each gene will be reported
If FALSE, a number of total length for all gene will be reported. Default is FALSE.}
}
\value{
a number or a vector
}
\description{
get gene length
}
