% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/procMaf.R
\name{Get_incon_mut}
\alias{Get_incon_mut}
\title{Get_incon_mut
Get Variants with incosistant gene annotation from
GSMuta reference data}
\usage{
Get_incon_mut(mutab, exome)
}
\arguments{
\item{mutab}{maf matrix}

\item{exome}{exome file.}
}
\value{
Get location of variants with incosistant gene annotation from
GSMuta reference data
}
\description{
Get_incon_mut
Get Variants with incosistant gene annotation from
GSMuta reference data
}
\examples{
\dontrun{
Get_incon_mut(mutab,exome)
}

}
