% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{get_mu_hat_mle}
\alias{get_mu_hat_mle}
\title{get_mu_hat_mle}
\usage{
get_mu_hat_mle(y, disp, offset, cores, ZI = FALSE)
}
\arguments{
\item{y}{input y}

\item{disp}{disp}

\item{offset}{offset}

\item{cores}{number of threads}

\item{ZI}{A boolen. Default is FALSE which Negative binomial model will be used.
If TRUE, zero-inflated model will be used.}
}
\value{
mu
}
\description{
from phi estimate mu
}
