% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{get_muhat}
\alias{get_muhat}
\title{get data mu for each gene #####}
\usage{
get_muhat(y, offset)
}
\arguments{
\item{y}{input data}

\item{offset}{offset}
}
\value{
muhat
}
\description{
get data mu for each gene #####
}
