% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getEnsemblID}
\alias{getEnsemblID}
\title{getEnsemblID}
\usage{
getEnsemblID(geneID,
  geneInfof = "~/Data/GSMuta_data/GSMutaRData/extdata/ensembl_92_exon_pos.hg38.rda")
}
\arguments{
\item{geneID}{gene symbol}

\item{geneInfof}{store the info symbol and ensembl ID.}
}
\value{
a vector of ensemble ID
}
\description{
get ensembl ID from gene symbol
}
