% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getGeneBgProb}
\alias{getGeneBgProb}
\title{getGeneBgProb}
\usage{
getGeneBgProb(exomeGene, prob, consequences = c(1, 2, 3, 4, 5),
  gid = NULL)
}
\arguments{
\item{exomeGene}{data.frame}

\item{prob}{probability of each tri-nucleotides mutation context}

\item{consequences}{default is c(1,2,3,4,5) which are nonsil mutation.}

\item{gid}{Gene ID list.}
}
\value{
data.frame background gene muation probabilty without regression
}
\description{
get background gene muation probabilty without regression
}
