% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getGeneSymbol}
\alias{getGeneSymbol}
\title{getGeneSymbol}
\usage{
getGeneSymbol(ensemblID,
  geneInfof = "~/Data/GSMuta_data/GSMutaRData/extdata/ensembl_92_exon_pos.hg38.rda")
}
\arguments{
\item{ensemblID}{ensemble ID}

\item{geneInfof}{store the info symbol and ensembl ID.}
}
\value{
a vector of gene symbol
}
\description{
get gene symbol from ensembl ID
}
