% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getsubData}
\alias{getsubData}
\title{getsubData}
\usage{
getsubData(Data, sampleN = NULL, gid = NULL)
}
\arguments{
\item{Data}{MutSet}

\item{sampleN}{A list of sample names.}

\item{gid}{A list of gene id.}
}
\value{
MutSet a subset of orginal MutSet
}
\description{
get subset of orginal MutSet
}
