% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TMBpred.R
\name{getTMBs}
\alias{getTMBs}
\title{getTMBs}
\usage{
getTMBs(y, offset, phi, cores, zero_p = 0, prior = NULL, span = 10,
  low = 10^(-4), high = 10^4)
}
\arguments{
\item{y}{A matrix for observed mutation count}

\item{offset}{offset}

\item{phi}{gene dispersion}

\item{cores}{A number which specifiy the nubmer of core to use for parallel computing.}

\item{zero_p}{gene zero fraction}

\item{prior}{prior parameters. If it is NULL, no prior will be used.}

\item{span}{The extension of low and high limit. High = high * span and low = low/span}

\item{low}{The lower limit.}

\item{high}{The higher limit.}
}
\value{
A data.frame contain the result.
}
\description{
predict TMB stat function
}
\examples{
\dontrun{
getTMBs(Data)
}

}
