% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/procMaf.R
\name{maf_dnp_converter}
\alias{maf_dnp_converter}
\title{maf_dnp_converter}
\usage{
maf_dnp_converter(mutab)
}
\arguments{
\item{mutab}{maf matrix}
}
\value{
maf file with converted SNP
}
\description{
maf_dnp_converter
}
\examples{
\dontrun{
maf_dnp_converter(maf)
}

}
