% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/procMaf.R
\name{mafCleanup}
\alias{mafCleanup}
\title{maf.cleanup}
\usage{
mafCleanup(maf, extraCols = NULL, keepNoncoding = FALSE,
  save = FALSE, fn = "./output.tsv", nonSilTypes = NULL,
  SilTypes = NULL, ncTypes = NULL, gid = NULL)
}
\arguments{
\item{maf}{Input maf file. It can be a path to either rda or tsv file or data.frame}

\item{extraCols}{Default is NULL which no extra column will be included. If "all", all columns
will be output. If an array, only specificed columns will be reported.}

\item{keepNoncoding}{A boolen. If TRUE, the noncoding variants will be keeped. Default is FALSE,
which noncoding variants will be removed.}

\item{save}{A boolen. If TRUE, the file will be saved, FALSE, the file will not be saved}

\item{fn}{string, the output file name}

\item{nonSilTypes}{To specify what variant classification should be considered as nonsilent variants.
Default is NULL which the default list of classification will be used.}

\item{SilTypes}{To specify what variant classification should be considered as silent variants.
Default is NULL which the default list of classification will be used.}

\item{ncTypes}{To specify what variant classification should be considered as variants in noncoding regions.
Default is NULL which the default list of classification will be used.}

\item{gid}{Gene ID list.}
}
\value{
cleaned up maf file.
}
\description{
maf.cleanup
}
\examples{
\dontrun{
maf.cleanup(x, output)
}

}
