% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecTMB.R
\docType{class}
\name{MutSet}
\alias{MutSet}
\title{MutSet class}
\format{An \code{\link{R6Class}} object.}
\usage{
MutSet
}
\value{
Object of class \code{MutSet}
}
\description{
MutSet class
}
\section{Slots}{

\describe{
\item{\code{mut}}{table from MAF}

\item{\code{gid}}{list of gene id for analysis}

\item{\code{exome}}{table contains all possible mutations}

\item{\code{exomeGene}}{table contains all posiible mutations for each gene.}

\item{\code{covar}}{table contains covariates for each gene}

\item{\code{mutContext}}{The context of 96 mutations}

\item{\code{samples}}{data.frame store sample info. if BED colunm is specified,
the background process will be using the regions limited to bed file.}

\item{\code{incosistantAnno}}{Variants with incosistant gene annotation from
GSMuta reference data .
\describe{
  \item{\code{new()}}{Create a MutSet }
  \item{\code{nonsilent()}}{Get nonsilent mutations table}
  \item{\code{silent()}}{Get silent mutations table}
  \item{\code{nc()}}{Get noncoding mutations table}
}}
}}

\examples{
\dontrun{
MutSet$new()
}
}
\keyword{datasets}
