% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{OverLap}
\alias{OverLap}
\title{OverLap}
\usage{
OverLap(Bed, regions, base0 = FALSE)
}
\arguments{
\item{Bed}{targeted bed file. data.frame contain colunms "Chromosome" "Start" "End"}

\item{regions}{path to bed file or data.frame}

\item{base0}{A boolen if it is 0 based.}
}
\value{
a list of row number which targetBed overlap with regions.
}
\description{
OverLap
}
\examples{
\dontrun{
  exomebed        = data.frame(Chromosome = exome[, "Chromosome"],
  Start = exome[, "pos"],
  End = exome[,"pos"])
  OverLap(exomebed, regions)
}

}
