% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TMBpred.R
\name{pred_TMB}
\alias{pred_TMB}
\title{pred_TMB}
\usage{
pred_TMB(x, params, WES = NULL, prior = list(mu = prior_bs$mu, sigma =
  prior_bs$sigma, lambda = prior_bs$lambda), mut.nonsil = FALSE,
  gid_nonsil_p = NULL, method = "MLE", cores = 10, span = 1,
  low = 10^(-4), high = 10^4)
}
\arguments{
\item{x}{test mutSet object}

\item{params}{A list contain result from fit}

\item{WES}{test mutSet object which contain whole exome data which will be used to generate the truth.}

\item{prior}{A list contain prior parameters}

\item{mut.nonsil}{A boolen. If it is TRUE, the nonsilent mutation will be used for prediction}

\item{gid_nonsil_p}{A vector of gid for passenger genes}

\item{method}{A string which specify the method to predict the TMB}

\item{cores}{A number which specifiy the nubmer of core to use for parallel computing.}

\item{span}{The extension of low and high limit. High = high * span and low = low/span}

\item{low}{The lower limit.}

\item{high}{The higher limit.}
}
\value{
A data.frame contain the result.
}
\description{
fit the data distribution to a mixture of Gaussian mixture
}
\examples{
\dontrun{
pred_TMB(Data)
}

}
