% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecTMB.R
\name{readData}
\alias{readData}
\title{read and load maf file and necessary reference file}
\usage{
readData(mutf, exomef, covarf, mutContextf, ref, samplef = NULL)
}
\arguments{
\item{mutf}{Path to maf file. Detail see \code{mafCleanup}.}

\item{exomef}{Path to exome file}

\item{covarf}{Path to covariable file}

\item{mutContextf}{Path to mutContext file.}

\item{ref}{Path to reference genome}

\item{samplef}{Path to sample file or data.frame. The column name to specify bed file must be BED}
}
\value{
a MutSet
}
\description{
read and load maf file and necessary reference file
}
