% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TMBpred.R
\name{remove_outliers}
\alias{remove_outliers}
\title{remove_outliers}
\usage{
remove_outliers(x, na.rm = TRUE, ...)
}
\arguments{
\item{x}{number of mutation}

\item{na.rm}{how to deal with na}

\item{...}{The extra parameters for quantile.}
}
\value{
outlier is labeled as NA
}
\description{
remove outlier based on 1.5 quantile
}
\examples{
\dontrun{
remove_outliers(x)
}

}
