% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/procMaf.R
\name{retrieve_context}
\alias{retrieve_context}
\title{retrieve_context}
\usage{
retrieve_context(mutab, ref)
}
\arguments{
\item{mutab}{maf matrix}

\item{ref}{genome reference path.}
}
\value{
Retrieve mutation context code for each variants
}
\description{
retrieve_context
}
\examples{
\dontrun{
maf_dnp_converter(mutab, ref, codeRetrieve)
}

}
