% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{zipMLE}
\alias{zipMLE}
\title{Calculate mu and zero_p by MLE for ZIP model #}
\usage{
zipMLE(x, offset = 1, tol = 1e-09)
}
\arguments{
\item{x}{observed mutation}

\item{offset}{offset.}

\item{tol}{tol of convergency}
}
\value{
a List of parameters
}
\description{
Calculate mu and zero_p by MLE for ZIP model #
}
\examples{
\dontrun{
zipMLE(Data, offset)
}

}
